Pick up xrt library from install
Introduced XILINX_XRT to override XILINX_SDX when looking for HAL
level xrt library.  If XILINX_XRT is defined, then XILINX_SDX will not
be searched.  If XILINX_XRT is not defined, then runtime will work as
current 2018.2.  This is subject to change shortly once everyting is
up and running.