xrtdeps.sh - allow use of arbitrary sudo program
Xilinx internal servers frequently use /tools/xgs/bin/sudo rather than
a version of the tool on the normal system path. This change allows
an environment variable called $SUDO to be used to point to the tool,
otherwise if unset it reverts to the simple bare `sudo` command as
before.