Support CDMA base addresses per feature rom
Number of changes to support CDMA as a shared resource ignored by user context.  Ultimately validation of exec buffer must ignore CUs that identify CDMA kernels.

To support this, the context now tracks regular CUs separately from CDMA kernels.  The CDMA kernels, which are used as regular CUs, are added to the context when the scheduler configuration assigns them a CU idx.