Enhancement: Change JSON output format from a multiset to an array. (#1866)
Description
----------
The current JSON output of "collections" is represented as a list.  Being that each list object has a key and the keys are duplicates of each other, this was causing issues with various JSON parser for this data was being placed into a set as opposed to a multset.

To help address this issue, instead of creating a list of objects, xclbinutil will now create an array of objects.

Note
----
When reading in the JSON file, both lists and arrays are supported.