Bug Fix: Added a command argument preprocessor implicit arguments (#2810)
* Bug Fix: Added a command argument preprocessor implicit arguments
This is a kludge fix to address changes in boost with regards on how it process implicit values.  In the past, adjacent key value pairs need not have an equal '=' value.  Going forward it now requires it.
For example:
--info myfile
was valued with some versions of boost.  With new versions of the boot libraries only the syntax:
--info=myfile
is valued.

* Spelling corrections.