Xrt kernel (#2893)
* Support XRT native kernel execution APIs

This is WIP.  This PR is functionally working to the point where
low level tests can be changed.

The kernel API implementation is delivered in libxrt_coreutils.  Many
changes in this PR are to facililate shim level calls from core/common
without link-time dependencies on libxrt_core.

* Fix windows build

* Fix edge compilation plus major cleanup

Cleanup is prep to facilitate implematation of shared execution API

* Fix windows build

* Fix system singleton to work with static linking

Fix bug in shim destruction that caused device to be closed before
bo_cache was destructed.  Since bo_cache dtor unmaps and releases its
exec BOs the cache must be deleted before the device is closed.

* Fix compile error after merge