CR-1063652 System hang on multi-threaded host accessing multiple CUs (#3326)
* CR-1063652 System hang on multi-threaded host accessing multiple CUs

Dynamically size the ERT command queue slots.  The default number of
slots in ERT (16), is not necessarily enough to keep > 16 CUs running
if there is dependency between the CUs.

In the CR, there are total of 30 CUs with some K-K streaming
dependency between CUs, e.g. one CU produces output for another CU.
When all consumers are running but no producers, then things stall.

The sizing of CQ is based on number of CUs in xclbin such that desired
number of slots is twice the CUs in the xclbin.  The computed number
of slots is bounded by the register map size of largest CU and finally
rounded to upper 16, 32, 64, 96, 128 to keep slots at register
boundary (16 to keep changes to a minimum).