Fix overload bugs (#4108)
Regression from #4098 and #4076.

std::forward of lvalue is lvalue.  xrt::bo overload must 
Variadic template overload behavior is surprising.  The xrt::bo
overload of set_arg was not called as expected.  Had to change from
(const xrt::bo&) to (xrt::bo&).  Tried (xrt::bo&&) which seemed a
better fit for the std::forward call in the variadic template, but
that doesn't match.