CR-1080628 Set kernel argument with sub buffer fails if xclbin multiple kernels (#4280)
Using sub-buffers with deferred allocation (until clSetKernelArg) was
broken since inception of sub buffer support in XRT.

When a buffer is allocated via set kernel arg, the CU connectivity is
used to determine where the buffer should be allocated.  If the buffer
is a sub-buffer, it is the parent buffer than must be allocated.  But
since the sub-buffer was used to look up connectivity, only this
buffer has the has the memory bank index.  So this PR contains fixes
that will pass the computed bank index to the parent buffer when it is
allocated.