WIP: Multiple xclbin support, add xclOpenContext from slot and cuname
As part of supporting multiple xclbins, compute unit context needs to
be opened by slot index and name.  This allows resource assignment to
reject the context request in specified slot and maybe load the xclbin
into a different slot.

The compute unit index is not guaranteed to be valid until successful
context creation.  This is a significant departure from current XRT
where cu indices are created as part of loading a single xclbin.