Fix race in bo and device construction (#7361)
* Fix race on xrt::bo construction

Multiple threads competing to construct the very first xrt::bo object
can hit a race in a device query for nodma check.

The check is cached in a member variable and this variable is
subsequently checked.  The check must be under a mutex lock.

* Fix race in core device creation

Add missing mutex for device map access.

Multiple threads can call compete in construction of shim device
objects.  This in turn constructs a core device object and inserts it
into a static map in coreutil.  This map needs to be protected not
only for insertion but also for reading.